refactor: move resource server code
In order to write new e2e test for the electron app, specifically, for the automated checks view, we need a way to "mock" axe-android calls and results. We are going to use a similar approach as `yarn with:mock-axe-andoid`, whichs spins up a local static file server to respond with a result file (in the future, we may need to get more sofisticated and have different results file for different scenarios but for now, one file should do).

For this to work properly, we need to generalize how we are spinning that resouces server. Adding a config type seems to fit this nicely as it allows to pack the port and the path we're currently passing ti the resource server under browser extension e2e scenario.
