+++ src/tests/miscellaneous/test-resource-server/resource-server-config.ts
@@ -1,4 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
-export const port = 9050;
+export type ResourceServerConfig = {
+    port: number;
+    path: string;
+};
